refactor sonar build config
